# core-banking-payments-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-baking-payments
  description: Payments Core Application
  contact:
    name: Firefly Software Foundation
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.getfirefly.io/payments
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  /api/v1/payment-orders/{paymentOrderId}/providers/{paymentProviderId}:
    get:
      tags:
        - Payment Providers
      summary: Get Payment Provider by ID
      description: Retrieve a specific payment provider by its unique identifier, within the context of a payment order.
      operationId: getPaymentProviderById
      parameters:
        - name: paymentOrderId
          in: path
          description: Unique identifier of the payment order
          required: true
          schema:
            type: string
            format: uuid
        - name: paymentProviderId
          in: path
          description: Unique identifier of the payment provider
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successfully retrieved the payment provider
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentProviderDTO'
        '404':
          description: Payment provider not found
    put:
      tags:
        - Payment Providers
      summary: Update Payment Provider
      description: Update an existing payment provider by its unique identifier.
      operationId: updatePaymentProvider
      parameters:
        - name: paymentOrderId
          in: path
          description: Unique identifier of the payment order
          required: true
          schema:
            type: string
            format: uuid
        - name: paymentProviderId
          in: path
          description: Unique identifier of the payment provider to update
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentProviderDTO'
      responses:
        '200':
          description: Payment provider updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentProviderDTO'
        '404':
          description: Payment provider not found
    delete:
      tags:
        - Payment Providers
      summary: Delete Payment Provider
      description: Remove an existing payment provider by its unique identifier, within the context of a payment order.
      operationId: deletePaymentProvider
      parameters:
        - name: paymentOrderId
          in: path
          description: Unique identifier of the payment order
          required: true
          schema:
            type: string
            format: uuid
        - name: paymentProviderId
          in: path
          description: Unique identifier of the payment provider to delete
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Payment provider deleted successfully
        '404':
          description: Payment provider not found
...
  /api/v1/payment-orders/{paymentOrderId}/providers:
    get:
      tags:
        - Payment Providers
      summary: Retrieve Payment Providers
      description: Retrieves a paginated list of payment providers for a specific payment order, with optional filtering.
      operationId: getAllPaymentProviders
      parameters:
        - name: paymentOrderId
          in: path
          description: Unique identifier of the payment order
          required: true
          schema:
            type: string
            format: uuid
        - name: pagination.pageNumber
          in: query
          description: Page number (0-based)
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: pagination.pageSize
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            format: int32
            default: 10
        - name: pagination.sortBy
          in: query
          description: Field to sort by
          required: false
          schema:
            type: string
        - name: pagination.sortDirection
          in: query
          description: Sort direction (ASC or DESC)
          required: false
          schema:
            type: string
            default: DESC
        - name: options.caseInsensitiveStrings
          in: query
          description: Enable case-insensitive string filtering
          required: false
          schema:
            type: boolean
            default: false
        - name: options.includeInheritedFields
          in: query
          description: Include fields from parent classes
          required: false
          schema:
            type: boolean
            default: false
        - name: filters.paymentOrderId
          in: query
          description: Exact filter for payment order id
          required: false
          schema:
            type: string
        - name: filters.providerName
          in: query
          description: Exact filter for provider name
          required: false
          schema:
            type: string
        - name: filters.externalReference
          in: query
          description: Exact filter for external reference
          required: false
          schema:
            type: string
        - name: filters.status
          in: query
          description: Exact filter for status
          required: false
          schema:
            type: string
            enum:
              - ACTIVE
              - INACTIVE
              - PENDING
              - SUSPENDED
        - name: filters.providerType
          in: query
          description: Exact filter for provider type
          required: false
          schema:
            type: string
        - name: filters.providerUrl
          in: query
          description: Exact filter for provider url
          required: false
          schema:
            type: string
        - name: filters.providerApiKey
          in: query
          description: Exact filter for provider api key
          required: false
          schema:
            type: string
        - name: filters.providerUsername
          in: query
          description: Exact filter for provider username
          required: false
          schema:
            type: string
        - name: filters.providerFee
          in: query
          description: Exact filter for provider fee
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[providerFee].from
          in: query
          description: Filter provider fee from value
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[providerFee].to
          in: query
          description: Filter provider fee to value
          required: false
          schema:
            type: string
        - name: filters.providerFeeCurrencyCode
          in: query
          description: Exact filter for provider fee currency code
          required: false
          schema:
            type: string
        - name: filters.providerResponseCode
          in: query
          description: Exact filter for provider response code
          required: false
          schema:
            type: string
        - name: filters.providerResponseMessage
          in: query
          description: Exact filter for provider response message
          required: false
          schema:
            type: string
        - name: filters.dateCreated
          in: query
          description: Exact filter for date created
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[dateCreated].from
          in: query
          description: Filter date created from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[dateCreated].to
          in: query
          description: Filter date created to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.dateUpdated
          in: query
          description: Exact filter for date updated
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[dateUpdated].from
          in: query
          description: Filter date updated from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[dateUpdated].to
          in: query
          description: Filter date updated to value
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Successfully retrieved payment providers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponse'
        '404':
          description: No payment providers found
    post:
      tags:
        - Payment Providers
      summary: Create Payment Provider
      description: Creates a new payment provider for the specified payment order.
      operationId: createPaymentProvider
      parameters:
        - name: paymentOrderId
          in: path
          description: Unique identifier of the payment order
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentProviderDTO'
      responses:
        '201':
          description: Payment provider created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentProviderDTO'
        '400':
          description: Invalid payment provider data provided
...
components:
  schemas:
...
    PaymentProviderDTO:
      required:
        - paymentOrderId
        - providerName
        - providerType
        - status
      type: object
      properties:
        dateCreated:
          type: string
          format: date-time
          readOnly: true
        dateUpdated:
          type: string
          format: date-time
          readOnly: true
        paymentProviderId:
          type: string
          format: uuid
          readOnly: true
        paymentOrderId:
          type: string
          format: uuid
        providerName:
          maxLength: 255
          minLength: 0
          type: string
        externalReference:
          maxLength: 100
          minLength: 0
          type: string
        status:
          type: string
          enum:
            - ACTIVE
            - INACTIVE
            - PENDING
            - SUSPENDED
        providerType:
          maxLength: 50
          minLength: 0
          type: string
        providerUrl:
          maxLength: 500
          minLength: 0
          pattern: ^https?://.*
          type: string
        providerApiKey:
          maxLength: 255
          minLength: 0
          type: string
        providerUsername:
          maxLength: 100
          minLength: 0
          type: string
        providerAccountId:
          maxLength: 100
          minLength: 0
          type: string
        providerFee:
          minimum: 0
          exclusiveMinimum: false
          type: number
        providerFeeCurrencyCode:
          pattern: ^[A-Z]{3}$
          type: string
        providerResponseCode:
          maxLength: 20
          minLength: 0
          type: string
        providerResponseMessage:
          maxLength: 500
          minLength: 0
          type: string
        providerTransactionId:
          maxLength: 100
          minLength: 0
          type: string
...
    PaginationResponse:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            type: object
            description: The list of items for the current page.
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...